+++ pruebas/01-reading-list/tonysantana1492/src/App.tsx
@@ -1,41 +1,21 @@
-import { useCallback, useEffect } from 'react'
+import { useEffect } from 'react'
 
 import { Books, Favorites, Loading, SideMenu } from './components'
-import { useBook, useLocalStore } from './hooks'
+import { useBook } from './hooks'
 import { useMenu } from './store'
 
 export const App = () => {
-  const { favorites, getBooks, updateFavorites: setFavoritesFromLocalStore, loading, setLoading } = useBook()
-  const { getFromLocalStorage } = useLocalStore()
+  const { favorites, getBooks, loading, updateFromLocalStorage } = useBook()
   const { isOpen } = useMenu()
 
-  const updateFromLocalStorage = useCallback(() => {
-    const favorites = getFromLocalStorage()
-    setFavoritesFromLocalStore({ updatedFavorites: favorites })
-  }, [getFromLocalStorage, setFavoritesFromLocalStore])
-
-  const handlegetBooks = useCallback(async () => {
-    setLoading(true)
-
-    try {
-      await getBooks()
-      updateFromLocalStorage()
-    } catch (error) {
-      console.error(error)
-    } finally {
-      setLoading(false)
-    }
-  }, [getBooks, setLoading, updateFromLocalStorage])
-
   useEffect(() => {
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    handlegetBooks()
-  }, [handlegetBooks])
+    getBooks()
+  }, [getBooks])
 
   useEffect(() => {
     window.addEventListener('storage', updateFromLocalStorage)
     return () => { window.removeEventListener('storage', updateFromLocalStorage) }
-  }, [setFavoritesFromLocalStore, getFromLocalStorage, updateFromLocalStorage])
+  }, [updateFromLocalStorage])
 
   if (loading) {
     return <Loading />
@@ -44,7 +24,7 @@ export const App = () => {
   return (
 		<main className="flex flex-row pt-4 h-screen">
 			<Books />
-			<section className="hidden lg:flex">{favorites.length > 0 && <Favorites />}</section>
+			<aside className="hidden lg:flex">{favorites.length > 0 && <Favorites />}</aside>
 			{isOpen && <SideMenu />}
 		</main>
   )
