+++ pruebas/01-reading-list/tonysantana1492/src/components/PageFilter.tsx
@@ -16,7 +16,7 @@ export const PageFilter: React.FC<Props> = ({ pageFilter, maxPages, updatePageFi
   return (
 		<div className="flex flex-col justify-center items-center w-2/4">
 			<div className='flex justify-between w-full items-center mb-3'>
-				<label className='text-xs font-bold text-gray-500' htmlFor={idRange}>Pages:</label>
+				<label className='text-xs font-bold text-gray-500' htmlFor={idRange}>Max pages:</label>
 				<p className='text-xs font-bold text-white bg-slate-900 px-2 rounded-full'>{pageFilter}</p>
 			</div>
 			<input role='filter-page' className='w-full h-2 bg-gray-300 rounded-lg appearance-none cursor-pointer accent-gray-800' id={idRange} type="range" onChange={handleChange} value={pageFilter} min={0} max={maxPages} />
